feat: get list of available issues
THe main logic will be in downloading single issue.

Maybe in the future this should be coverted so that multiple issues can
be downloaded at the same time, but for now it will be only single
issue. Issues contains multiple articles which are in separated files so
those would be dowloaded concurrently